requirements.txt
# ============================================================================
# Multi-Agent Research Intelligence Platform - Dependencies
# ============================================================================

# Core LangChain & LangGraph
langgraph==1.0.0
langchain==1.0.0
langchain-groq==1.0.0
langchain-huggingface==1.0.0
langchain-chroma==1.0.0
langchain-core==1.0.0
langchain-text-splitters==1.0.0

# Search & Data Sources
arxiv==2.2.0
google-search-results==2.4.2  # serpapi
requests==2.32.5

# Vector DB & Embeddings
chromadb==1.2.0
sentence-transformers==5.1.1

# Machine Learning
scikit-learn>=1.5.0
numpy>=1.26.4

# Visualization
matplotlib>=3.8.0
networkx>=3.0
plotly>=5.0.0

# Web UI
streamlit>=1.40.0
graphviz>=0.20.1
pyvis>=0.3.2
plotly>=5.0.0
Pillow>=10.0.0

# Utilities
python-dotenv==1.0.1
beautifulsoup4==4.12.3

# Apache Airflow (Optional - for orchestration)
# Uncomment if using Airflow mode
# apache-airflow==2.7.3
